layoutlmv3_ner/Lec_5_7/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# OCR相关
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# 文档处理
pdf2image>=1.16.0  # Fixed: Changed from 3.1.0 to 1.16.0
Pillow>=9.5.0
opencv-python>=4.8.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 标注工具
label-studio>=1.8.0
label-studio-sdk>=0.0.30

# 其他工具
tqdm>=4.65.0
requests>=2.31.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0